Separated AuthenticatedForm into three files: Welcome, Login, Signup
