+++ web/package.json
@@ -6,6 +6,7 @@
     "@apollo/client": "^3.4.8",
     "@cloudinary/base": "^1.0.0-beta.4",
     "@cloudinary/react": "^1.0.0-beta.8",
+    "@material-ui/core": "^4.12.3",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
