+++ web/src/App.tsx
@@ -1,5 +1,7 @@
 import React, { useMemo } from "react";
-import { useCookies } from "react-cookie";
+// import { ThemeProvider } from "@material-ui/styles";
+// import { createTheme, responsiveFontSizes } from "@material-ui/core/styles";
+import { StylesProvider } from "@material-ui/core/styles";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import {
 	ApolloClient,
@@ -40,32 +42,59 @@ function App() {
 		};
 	});
 
+	// const theme = useMemo(() => {
+	// 	return responsiveFontSizes(
+	// 		createTheme({
+	// 			palette: {
+	// 				primary: {
+	// 					main: "#00a6de",
+	// 				},
+	// 			},
+	// 			typography: {
+	// 				fontFamily: "Circular",
+	// 				fontWeightLight: 300,
+	// 				fontWeightRegular: 500,
+	// 				fontWeightMedium: 600,
+	// 				fontWeightBold: 700,
+	// 			},
+	// 			props: {
+	// 				MuiInput: {
+	// 					disableUnderline: true,
+	// 				},
+	// 			},
+	// 		})
+	// 	);
+	// }, []);
+
 	const client = useMemo(() => {
 		return new ApolloClient({
 			link: authLink.concat(httpLink),
 			cache: new InMemoryCache(),
-		})
+		});
 	}, [authLink]);
 
 	return (
 		<ApolloProvider client={client}>
-			<Router>
-				<AppState>
-					<Navbar />
-
-					<Switch>
-						<Route exact path={LANDING} component={Landing} />
-						<Route path={LISTINGS} component={Listings} />
-						<Route
-							path={LISTING}
-							render={(renderProps) => (
-								<Listing id={renderProps.match.params.id} />
-							)}
-						/>
-						<Route exact path="/users" component={Users} />
-					</Switch>
-				</AppState>
-			</Router>
+			{/* <ThemeProvider theme={theme}> */}
+			<StylesProvider injectFirst>
+				<Router>
+					<AppState>
+						<Navbar />
+						<Switch>
+							<Route exact path={LANDING} component={Landing} />
+							<Route path={LISTINGS} component={Listings} />
+							<Route
+								path={LISTING}
+								render={(renderProps) => (
+									<Listing id={renderProps.match.params.id} />
+								)}
+							/>
+							<Route exact path="/users" component={Users} />
+						</Switch>
+					</AppState>
+				</Router>
+			</StylesProvider>
+			{/* </ThemeProvider> */}
 		</ApolloProvider>
 	);
 }
