+++ web/src/stylesheets/components/FormError.scss
@@ -1,0 +1,8 @@
+.FormError {
+  padding-top: 5px;
+  display: flex;
+
+  span {
+    margin-right: 3px;
+  }
+}
