+++ web/src/stylesheets/main.scss
@@ -1,32 +1,22 @@
 @charset "UTF-8";
 
 // 1. Configuration and helpers
-@import
-  'abstracts/variables',
-  'abstracts/mixins';
+@import "abstracts/variables", "abstracts/mixins";
 
 // 2. Vendors
-@import
-  'vendors/normalize';
+@import "vendors/normalize";
 
 // 3. Base stuff
-@import
-  'base/base',
-  'base/fonts',
-  'base/typography',
-  'base/animations';
+@import "base/base", "base/fonts", "base/typography", "base/animations";
 
 // 4. Layout-related sections
-@import
-  'layout/header',
-  'layout/footer';
+@import "layout/header", "layout/footer";
 
 // 5. Components
-@import
-  'components/Loading';
+@import "components/Loading", "components/AuthenticatedForm",
+	"components/FormError.scss";
 //   'components/title',
 //   'components/hamburger';
 
 // 6. Page-specific styles
-@import
-  'pages/Listing'
+@import "pages/Listing";
