--- web/src/utils/age.tsx
@@ -1,12 +1,0 @@
-const getAge = (DOB: string) => {
-	let today = new Date();
-	let birthDate = new Date(DOB);
-	let age = today.getFullYear() - birthDate.getFullYear();
-	let m = today.getMonth() - birthDate.getMonth();
-	if (m < 0 || (m === 0 && today.getDate() < birthDate.getDate())) {
-		age--;
-	}
-	return age;
-};
-
-export { getAge };
