+++ web/src/utils/cookies.tsx
@@ -1,4 +1,4 @@
-function getCookie(name: string) {
+const getCookie = (name: string) => {
 	const value = `; ${document.cookie}`;
 	const parts = value.split(`; ${name}=`);
 	if (parts.length === 2) return parts?.pop()?.split(";").shift();
