+++ web/src/utils/regex.tsx
@@ -1,0 +1,5 @@
+const symbolRegex = /[-!@#$%^&*()_+|~=`{}\[\]:";'<>?,./]/;
+const emailRegex =
+	/^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+
+export { symbolRegex, emailRegex };
